requirements.txt
fastapi[standard]
pydantic
psycopg[binary]
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart